refactor: extract decision tree actions composable
